requirements.txt
# Core dependencies
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0

# NLP and Text Processing
transformers>=4.20.0
sentence-transformers>=2.2.0
tokenizers>=0.12.0
nltk>=3.7
spacy>=3.4.0

# Graph Libraries
networkx>=2.7.0
neo4j>=5.0.0
torch-geometric>=2.0.0  # For GFM encoder

# Vector Database and Search
faiss-cpu>=1.7.0  # Use faiss-gpu for GPU support
elasticsearch>=8.0.0
pymongo>=4.0.0

# Document Processing
PyPDF2>=2.0.0
pdfminer.six>=20200720
beautifulsoup4>=4.10.0

# LLM API
openai>=0.27.0  # For OpenAI API
anthropic>=0.3.0  # For Claude API
qianfan>=0.0.1  # For Baidu API

# Web and API
fastapi>=0.85.0
uvicorn>=0.18.0
requests>=2.28.0
aiohttp>=3.8.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
python-dotenv>=0.20.0
loguru>=0.6.0

# Test and Development
pytest>=7.0.0
black>=22.6.0
isort>=5.10.0
mypy>=0.960 